Fix UpdateProject function.
We were only updating the Tree model and not our internal project list. This was breaking the RemoveProjects function which relies on instances in both list to be identical as it uses the Remove function from List class. Make sure we also update our list with the new project description.