Fix lineup edit not updating team tagger.
Indeed when modifying the lineup event we need to reload like a substitution event. Also fix Reload method which was just reloading the teams but was not showing the lineup. In the normal scenario this update for the line up is coming from the CurrentTime being set to 0 but editing the lineup afterwards will never trigger that update anymore. We do it for each reload. This fixes LON-198.