Get accurate timestamps for frame drawings
The video sinks only provide the exact frame timestamp as the
current duration when it's in the PAUSED state.
This commit modifies the IPlayback interface to support
synchronous calls to play/pause/stop and calls pause
synchronously before getting the current position for the
frame that is going to be drawn.