Fix deserialization with circular references
Check if the object ID is in the stack and return this instance.
For example, when loading a Project from the DB with a TimelineEvent
referencing this Project, the Project instance is not yet
in the cache when the TimelinEvent.Project property is deserialized.
Withtout using the instance in the stack we would try to deserialized
this Project again since it's not in the Cache, entering an infinite loop.