Override all methods that needs to processed by the Loader Fody
The Loader Fody that injects the CheckIsLoaded method in all property
getters that are not-preloaded can only do it in properties of types
that are IStorable and parents, but not in child types, since those
types don't implement the IStorable interface and could be used
as base classes for no-IStorable types.
By making these methods virtual in the base class and overriding
them in TimelineEvent to warranty that the getters will be injected
correctly.