Fix regression in the teams and dashboards manager
When a template is saved, the model needs to be updated with
the saved template since it's a different object reference
to prevent unsave changes to be saved in the loaded template.
But in the same way, when the template is saved, the reference
in the model needs to be updated too.